Audio Functionality, Bug Fixes & Optimisations
# Main Menu
- Scaled Main Menu Scene to 1920x1080 (previously much lower I don't remember now) for consistency across scenes.
- Renamed "MainMenu.cs" to "SceneHandler.cs".
- Generalised all functions previously present in "MainMenu.cs" to a better Scene Management approach.

# Game Over & Pause Panels
- Removed "GameOver.cs" due to it being a redundant script. All of its functions are already covered in "SceneHandler.cs".
- Removed "KillerTrap.cs" due to its functionality being optimally handled in Player Collisions instead of a new instance of the script per each trap.
- Added Trap Killing functionality in PlayerCollisions.cs
.
# Background
- Added Background canvas and image.

# Cleanup
- Removed multiple instances of missing script components being needlessly attached to objects.
- Removed multiple instances of needless objects never being enabled/used.

# Audio
- Added Game Over audio functionality through the AudioManager.

# UI/HUD
- Fixed UI/HUD not showing previously setup CrystalCounter.
- Added working level background.
- Added Audio Settings Panel to the Pause Menu.

# Null checks
- Added checks for SerializeField objects and GetComponent returned variables in all scripts.

# Known Remaining Bugs:
- The "Resume" button in the Pause Menu does not work. Escape key works, but using the button does not.
- Pause Menu UI and Main Menu UI need tuning after scaling.

+++ Assets/Scripts/DataPersistenceManager.cs
@@ -14,18 +14,30 @@ public class DataPersistenceManager {
 	}
 
 	static public int loadInt(string key) {
+		if (!PlayerPrefs.HasKey(key)) {
+			return -1;
+		}
 		return PlayerPrefs.GetInt(key);
 	}
 
 	static public float loadFloat(string key) {
+		if (!PlayerPrefs.HasKey(key)) {
+			return -1f;
+		}
 		return PlayerPrefs.GetFloat(key);
 	}
 
 	static public string loadString(string key) {
+		if (!PlayerPrefs.HasKey(key)) {
+			return "";
+		}
 		return PlayerPrefs.GetString(key);
 	}
 
 	static public void erase(string key) {
+		if (!PlayerPrefs.HasKey(key)) {
+			return;
+		}
 		PlayerPrefs.DeleteKey(key);
 	}
